Allow hindent.el to work with older hindent versions
With this change, hindent-style continues to default to nil, but when it
is set, it will be passed to hindent.

This allows the elisp to work with older hindent programs, and if there
is a mismatch between the user's hindent program and local settings of
hindent-style, they will be made aware of it and able to resolve it.