Add --validate option, which checks that a file has been formatted
If a project wishes to standardise on hindent's formatting, it's
helpful to have a build step which will run hindent on the project's
Haskell files, and exit if any have not been correctly formatted.

This commit adds a --validate option for this purpose, in line with
the corresponding option in elm-format.