Add Dockerfile with hindent for easy use via Docker
There are Docker Hub hooks in the `hooks` folder to automatically build
and push a Docker image with the given hindent version (for example,
with a Docker Hub rule that triggers on Git tags).

The Dockerfile can be tested with

```
scripts/test-docker
```

The Bash scripts have been linted with
[ShellCheck](https://github.com/koalaman/shellcheck) as in

```
docker run --interactive --rm --volume "$(pwd):/mnt" \
  koalaman/shellcheck:v0.7.1 \
  hooks/build hooks/post_push scripts/test-docker
```

This addresses mihaimaruseac/hindent#571. A Docker Hub autobuild rule
would need to be configured separately.