Break a long type application (not `TypeApplications`)
Although it was reported as a type synonym issue, it was a general
problem with types being written, not limited to type synonyms.

This commit involves a considerable change. While we were able to keep
each line within 80 lines, honestly, I feel that some parts lost
readability (especially the multi-line headers in the instance
declarations).

Fixes: #534