Make `reformat` return a `ByteString` (#720)
* Make `reformat` return a `ByteString`

In the early days, `reformat` simply returned the `Builder` contained in
`PrintState`, thus it was reasonable for the return type of `reformat`
to be `Builder`. However, `reformat` started changing `Builder` to
`ByteString` once, manipulating it, and then back to `Builder` again, in
order to support prefixes and such. This was not very reasonable, and
instead, I thought it would be better to simply return `ByteString`
since it would be easier for the user to use it.

I am not familiar with the topic of lazy vs strict. The reason why I
made `reformat` return the strict version is that `reformat` accepts the
same type of value as an argument. However, since `Builder` can only
convert to lazy ByteString, perhaps it is more convenient to return the
lazy version.

* Add a changelog

* Fix an error