`hindent-mode` Various small changes (#1096)
* hindent.el: Show error on exit code ≠ 0

Instead of checking for just (= ret 1), we should really show an error
whenever the exit code is not clean (i.e., 0). This prevents
hindent-mode from silently failing, which gives the indication of a
command not doing anything at all.

* hindent.el: Fix bytecompile warnings

+ Enable lexical binding.
+ Replace cl-letf with cl-labels.
+ Depend on Emacs 27.1 for replace-region-contents, which is old enough
  at this point to have trickled down to even Debian stable.
+ Only require cl-lib at compile time.

* hindent.el/hindent-extra-args: Allow type to be a function

This is useful for backends (e.g., fourmolu) that allow looking for a
cabal file in some kind of directory. For example, utilising project.el,
one could write something like

    (setq hindent-process-path "fourmolu")
    (setq hindent-extra-args
          (lambda ()
            `("--stdin-input-file" (project-root (project-current)))))

* hindent.el: Update URL